dna: relocate drift detection to pkg/dna/drift (Issue #1757)
The full drift detection implementation — Detector interface, all data
types and enums, DetectorConfig, DefaultDetectorConfig(), NewDetector(),
and the detector struct with all methods — is relocated from
features/steward/dna/drift/ to the neutral pkg/dna/drift/ package.

This decouples the controller from steward-internal packages (parent
epic #1754): pkg/dna/drift has no steward-specific imports, only
api/proto/common, pkg/logging, and stdlib.

What changed:

- pkg/dna/drift/drift.go (new): canonical package with full
  implementation copied verbatim from features/steward/dna/drift/
  interfaces.go and detector.go

- pkg/dna/drift/drift_test.go (new): tests for DefaultDetectorConfig()
  returns non-nil and NewDetector(nil,nil) returns a non-nil Detector
  with nil error

- features/steward/dna/drift/interfaces.go: all type definitions
  replaced with type aliases (= pkgdrift.T) so existing steward callers
  compile without modification

- features/steward/dna/drift/detector.go: implementation removed;
  NewDetector and DefaultDetectorConfig are now thin wrappers
  delegating to pkg/dna/drift

- features/reports/**/*.go and *_test.go (11 files): drift import
  updated from features/steward/dna/drift to pkg/dna/drift; all
  referenced types (Detector, DriftEvent, DriftSeverity, severity
  constants) exist in the canonical package

go build ./pkg/dna/drift/ and go build ./features/steward/dna/drift/
both pass; grep for the old import string in features/reports/ returns
zero hits; make test-agent-complete passes.

Fixes #1757
